move stuff + address nit comments
